Update copyright and license information
Added note, inc. as a co-copyright holder along with Adobe, in both the LICENSE and COPYRIGHT files. Also introduced a NOTICE file to clarify the origin of the software and respective copyright holders for both the original work and the modifications. This ensures a clear distinction between the original work and the modifications made by note, inc.